pyproject.toml
[project]
name = "employee-tracking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi>=0.135.3",
    "mysql-connector-python>=9.6.0",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.44.0",
]
